[AutoDiff] Fix memory leaks caused by partial application handling.
In VJPEmitter, if the original function call has substitutions, we `partial_apply` it with no arguments to specialize it. This `partial_apply` is not being released. JVP and VJP are being specialized the same way, but they are not being released either.

To fix this, we release the `@differentiable` function returned by `autodiff_function`, which will release the original and the associated functions tht are to be filled in later altogether. If the original function does not have substitutions, we retain the original function to balance out the release of the `@differentiable` function that comes later. As a result, `ADContext::promoteToDifferentiableFunction` no longer needs to retain the associated functions.

Example where the original function has substitutions:
```
f' = partial_apply f<...>()
f_diff = autodiff_function f'
release_value f_diff
```

Example where the original function does not have substitutions:
```
retain_value f
f_diff = autodiff_function f
release_value f_diff
```

Note: This makes the `autodiff_function` folding optimization no longer able to detect the pattern, but it is necessary. We can rewrite the optimization later.

This should fix [TF-621](https://bugs.swift.org/browse/TF-621).